Forward the dry_run flag to EditableFM
This ensure that we don't write a file inside .dump()

+++ academic/editFM.py
@@ -6,20 +6,21 @@ yaml = YAML()
 
 
 class EditableFM:
-    def __init__(self, base_path: Path, delim: str = "---"):
-        self.base_path = Path(base_path)
+    def __init__(self, base_path: Path, delim: str = "---", dry_run: bool = False):
+        self.base_path = base_path
         if delim != "---":
             raise NotImplementedError("Currently, YAML is the only supported front-matter format.")
         self.delim = delim
         self.fm = []
         self.content = []
         self.path = ""
+        self.dry_run = dry_run
 
     def load(self, file: Path):
         self.fm = []
         self.content = []
         self.path = self.base_path / file
-        if not self.path.exists():
+        if self.dry_run and not self.path.exists():
             self.fm = dict()
             return
 
@@ -41,6 +42,8 @@ class EditableFM:
 
     def dump(self):
         assert self.path, "You need to `.load()` first."
+        if self.dry_run:
+            return
 
         with open(self.path, "w", encoding="utf-8") as f:
             f.write("{}\n".format(self.delim))
